services/ingest/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
python-arango>=7.9.0
pydantic>=2.5.0
structlog>=24.1.0
python-dotenv>=1.0.0

